+++ src/vite-env.d.ts
@@ -5,6 +5,11 @@ declare module "*.csv?raw" {
   export default content;
 }
 
+declare module "*.jpg?inline" {
+  const src: string;
+  export default src;
+}
+
 interface ImportMetaEnv {
   readonly VITE_SUPABASE_URL: string;
   readonly VITE_SUPABASE_ANON_KEY: string;